Tweak double-tap behavior.
Previously, double-tapping a dual-mode key would trigger its TAP action
on the downstroke of the second keypress.  However, that got to be very
obnoxious in the following situation:

- the f key exposes additional punctuation, so that f+u types a '('
- attempting to type any function name ending in an f, like 'printf('

The desired double tap behavior is distinguished from normal typing by
wanting a very large number of auto-repeats of the key in question, so
now the test for double taps is only triggered after the key in question
was held longer than HOLD_MS without another keypress.

This introduces a secondary delay before the autorepeats kick in under X
since autorepeats are ignored from the keyboard device and instead
calculated by X, and that timer starts at least HOLD_MS after the
physical keypress.  But this is much more tolerable than constant
interruptions to normal typing.