fix super-annoying double-shift bug
Turns out I was setting ev.code and reading it again immediately, but
assuming it was the old value I was reading.  This was causing a pedal
key which mapped to shift to make the real shift key appear as if it had
been released, when it wasn't.

But I also fixed a few oddities along the way.  For example, we now
dedup inputs going into a resolver, because otherwise two overlapping
keys would create a similar situation as described above.  Also, we now
dedup outputs for the entire sdiol application, because it presents to
the kernel as a single keyboard device, therefore it should not be
emitting multiple presses of a key without releases in between.